requirements.txt
Django
djangorestframework
python-dotenv
django-phonenumber-field
phonenumbers
mysqlclient
django-cors-headers